examples/streaming/python/requirements.txt
python-socketio[client]>=5.11
